+++ src/video_core/renderer_vulkan/vk_rasterizer.h
@@ -4,16 +4,12 @@
 
 #pragma once
 
-#include "common/vector_math.h"
 #include "core/hw/gpu.h"
 #include "video_core/rasterizer_accelerated.h"
-#include "video_core/regs_lighting.h"
-#include "video_core/regs_texturing.h"
 #include "video_core/renderer_vulkan/vk_pipeline_cache.h"
 #include "video_core/renderer_vulkan/vk_stream_buffer.h"
 #include "video_core/renderer_vulkan/vk_texture_runtime.h"
 #include "video_core/shader/shader.h"
-#include "video_core/shader/shader_uniforms.h"
 
 namespace Frontend {
 class EmuWindow;
@@ -125,12 +121,6 @@ private:
     /// Syncs the cull mode to match the PICA register
     void SyncCullMode();
 
-    /// Syncs the depth scale to match the PICA register
-    void SyncDepthScale();
-
-    /// Syncs the depth offset to match the PICA register
-    void SyncDepthOffset();
-
     /// Syncs the blend enabled status to match the PICA register
     void SyncBlendEnabled();
 
@@ -139,18 +129,6 @@ private:
 
     /// Syncs the blend color to match the PICA register
     void SyncBlendColor();
-
-    /// Syncs the fog states to match the PICA register
-    void SyncFogColor();
-
-    /// Sync the procedural texture noise configuration to match the PICA register
-    void SyncProcTexNoise();
-
-    /// Sync the procedural texture bias configuration to match the PICA register
-    void SyncProcTexBias();
-
-    /// Syncs the alpha test states to match the PICA register
-    void SyncAlphaTest();
 
     /// Syncs the logic op states to match the PICA register
     void SyncLogicOp();
@@ -169,46 +147,6 @@ private:
 
     /// Syncs the depth test states to match the PICA register
     void SyncDepthTest();
-
-    /// Syncs the TEV combiner color buffer to match the PICA register
-    void SyncCombinerColor();
-
-    /// Syncs the TEV constant color to match the PICA register
-    void SyncTevConstColor(std::size_t tev_index,
-                           const Pica::TexturingRegs::TevStageConfig& tev_stage);
-
-    /// Syncs the lighting global ambient color to match the PICA register
-    void SyncGlobalAmbient();
-
-    /// Syncs the specified light's specular 0 color to match the PICA register
-    void SyncLightSpecular0(int light_index);
-
-    /// Syncs the specified light's specular 1 color to match the PICA register
-    void SyncLightSpecular1(int light_index);
-
-    /// Syncs the specified light's diffuse color to match the PICA register
-    void SyncLightDiffuse(int light_index);
-
-    /// Syncs the specified light's ambient color to match the PICA register
-    void SyncLightAmbient(int light_index);
-
-    /// Syncs the specified light's position to match the PICA register
-    void SyncLightPosition(int light_index);
-
-    /// Syncs the specified spot light direcition to match the PICA register
-    void SyncLightSpotDirection(int light_index);
-
-    /// Syncs the specified light's distance attenuation bias to match the PICA register
-    void SyncLightDistanceAttenuationBias(int light_index);
-
-    /// Syncs the specified light's distance attenuation scale to match the PICA register
-    void SyncLightDistanceAttenuationScale(int light_index);
-
-    /// Syncs the shadow rendering bias to match the PICA register
-    void SyncShadowBias();
-
-    /// Syncs the shadow texture bias to match the PICA register
-    void SyncShadowTextureBias();
 
     /// Syncs and uploads the lighting, fog and proctex LUTs
     void SyncAndUploadLUTs();
@@ -283,19 +221,6 @@ private:
     Surface null_surface;
     Surface null_storage_surface;
 
-    struct {
-        Pica::Shader::UniformData data{};
-        std::array<bool, Pica::LightingRegs::NumLightingSampler> lighting_lut_dirty{};
-        bool lighting_lut_dirty_any = true;
-        bool fog_lut_dirty = true;
-        bool proctex_noise_lut_dirty = true;
-        bool proctex_color_map_dirty = true;
-        bool proctex_alpha_map_dirty = true;
-        bool proctex_lut_dirty = true;
-        bool proctex_diff_lut_dirty = true;
-        bool dirty = true;
-    } uniform_block_data = {};
-
     std::array<SamplerInfo, 3> texture_samplers;
     SamplerInfo texture_cube_sampler;
     std::unordered_map<SamplerInfo, vk::Sampler> samplers;
@@ -310,15 +235,6 @@ private:
     std::size_t uniform_buffer_alignment;
     std::size_t uniform_size_aligned_vs;
     std::size_t uniform_size_aligned_fs;
-
-    std::array<std::array<Common::Vec2f, 256>, Pica::LightingRegs::NumLightingSampler>
-        lighting_lut_data{};
-    std::array<Common::Vec2f, 128> fog_lut_data{};
-    std::array<Common::Vec2f, 128> proctex_noise_lut_data{};
-    std::array<Common::Vec2f, 128> proctex_color_map_data{};
-    std::array<Common::Vec2f, 128> proctex_alpha_map_data{};
-    std::array<Common::Vec4f, 256> proctex_lut_data{};
-    std::array<Common::Vec4f, 256> proctex_diff_lut_data{};
 };
 
 } // namespace Vulkan
